move common code to LocalCLI module and added the ability to execute scripts from local file
